packages/sutra-storage-client/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sutra-storage-client"
version = "1.0.0"
description = "gRPC client for Sutra Storage Server"
readme = "README.md"
authors = [
    {name = "Sutra AI Team", email = "team@sutra-ai.dev"}
]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
requires-python = ">=3.8"
dependencies = [
    "grpcio>=1.60.0",
    "protobuf>=4.25.0",
    "numpy>=1.24.0",
]

[project.urls]
Homepage = "https://github.com/sutra-ai/sutra-models"
Repository = "https://github.com/sutra-ai/sutra-models"

[tool.setuptools.packages.find]
where = ["."]
include = ["sutra_storage_client*"]

[tool.setuptools.package-data]
sutra_storage_client = ["*.py", "*_pb2.py", "*_pb2_grpc.py"]
